[ fix ] 서버리스 api 줄이기 위해 통합
